+++ tests/testthat/test-captions.R
@@ -50,7 +50,7 @@ test_that("html_document2 captions", {
          output_file = "caption.html", quiet = TRUE)
 
   doc <- read_html(file.path(dirname(rmd_file), "caption.html"))
-  captions_ids <- xml_find_all(doc, "//table/caption/span")
+  captions_ids <- xml_find_all(doc, "//table/caption//span")
   expect_equal(xml_attr(captions_ids, "id"), c("tab:table1", "tab:table2"))
   txt <- gsub("[\n\r]+", "", xml_text(xml_find_all(doc, "//table/caption")))
   expect_equal(txt, c("Table 1: a table caption", "Table 2: a table caption"))
@@ -101,7 +101,7 @@ test_that("html_document captions", {
   doc <- read_html(file.path(dirname(rmd_file), "caption.html"))
   captions <- xml_find_all(doc, "//table/caption")
   expect_equal(
-    sum(xml_text(captions) %in% "a table caption"),
+    sum(grepl("a table caption", xml_text(captions))),
     2
   )
 })